+++ driver/main.go
@@ -134,6 +134,10 @@ func run() int {
 
 	// load pipeline
 	pl = pipeline.New(*driverDir, *pluginDir, *configFile)
+	if *inputType == "socket" {
+		emptyType := ""
+		inputType = &emptyType
+	}
 	err := pl.Load(*inputType)
 	if err != nil {
 		logger.Error.Println("Unable to load pipeline error: ", err.Error())